remove todo, fix list cell bugs
